Move tests to a common file
